+++ docs/Whats-New.md
@@ -295,6 +295,7 @@ New:
 - Additions to automatic passenger deletion (by Starkku)
 - Buildings considered as vehicles (by Starkku)
 - `CanC4=false` building zero damage toggle (by Starkku)
+- OpenTopped transport target sharing customization (by Starkku)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)